Update API guard to allow public route access
Public routes are accessible to anyone but the guard only allows routes with the API key. The guard should be updated to allow public route access

Add control flow for testing if the route has a public decortor. Return true is the route is public, bypassing the need for the API key. We get isPublic from the reflector class.