Modularize the commons code
API guard uses the reflector and config modules, which means those modules need to be imported somewhere. By modularing the commons code, we allow easier access to the commons stuff while being able to bring in other modules.

Create a commons module with the APP_GUARD as a provider. Remove the API guard imports from main.ts and add the commons module as an import to the app module.